env vars added and some validation refactor

+++ backend/controllers/product.controller.js
@@ -34,7 +34,7 @@ export const updateProduct = async (req, res) => {
 
     const product = req.body
 
-    if( !mongoose.Types.ObjectId.isValid(id) ){
+    if( !mongoose.Types.ObjectId.isValid(id) ) {
         return res.status(404).json({ success:false, message:"Invalid Product Id" })
     }
 
@@ -49,11 +49,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const { id } = req.params
 
+    if( !mongoose.Types.ObjectId.isValid(id) ) {
+        return res.status(404).json({ success:false, message:"Invalid Product Id" })
+    }
+
     try {
         await Product.findByIdAndDelete(id)
         res.status(200).json({ success:true, message:"Product deleted" })
     } catch (error) {
         console.log("Error in delete product: ", error.message)
-        res.status(404).json({ success:false, message:"Product not found!" })
+        res.status(500).json({ success:false, message:"Server Error" })
     }
 }
